Make outside game components mobile friendly
