+++ packages/frontend/src/components/createGame/CreateGame.module.scss
@@ -9,7 +9,6 @@
     border: 1px solid black;
     border-radius: 5px;
     padding: 20px;
-    width: 600px;
     height: fit-content;
     background: vars.$white;
 }
