+++ packages/frontend/src/components/inGame/components/GoHome.module.scss
@@ -5,4 +5,5 @@
     top: 10px;
     left: 10px;
     background: vars.$white;
+    z-index: 1;
 }
